requirements.txt
azure-identity==1.13.0
azure-mgmt-containerservice==24.0.0
boto3==1.21.0
botocore==1.24.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
google-api-core==2.11.1
google-auth==2.19.1
google-cloud-container==2.23.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
Jinja2==3.1.2
kubernetes==26.1.0
MarkupSafe==2.1.3
oauthlib==3.2.2
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
websocket-client==1.5.3
azure-identity==1.13.0
azure-mgmt-containerservice==24.0.0